requirements.txt
apprise==1.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
idna==3.4
importlib-metadata==5.0.0
Markdown==3.4.1
mypy-extensions==0.4.3
oauthlib==3.2.2
pathspec==0.10.1
# pkg-resources==0.0.0
platformdirs==2.5.3
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.4.0
urllib3==1.26.12
zipp==3.10.0
google-auth
google-auth-oauthlib
google-api-python-client
drive-cli
